feat. 좌표가 두 개일 경우, 거리 출력. refactor. Coordinates일급 컬렉션 불변성 보장을 위해 getter 시 배열로 값을 반환

+++ src/main/java/Model/Coordinate.java
@@ -9,11 +9,11 @@ public class Coordinate {
         this.y= new CoordinateVariable(y);
     }
 
-    double getX(){
+    int getX(){
         return this.x.getValue();
     }
 
-    double getY(){
+    int getY(){
         return this.y.getValue();
     }
